requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
defusedxml==0.7.1
et_xmlfile==2.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
numpy==2.1.3
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
python-calamine==0.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.7
pytz==2024.2
pyxlsb==1.0.10
six==1.16.0
sniffio==1.3.1
timezones==3.0.0
tzdata==2024.2
xlrd==2.0.1
XlsxWriter==3.2.0
